Refactors parsers
